NXP-25526: move AuditQueryBuilder common code to nuxeo-core-query
